Refactor: create cleaer architecture
- create separate handlers
- try to create models
- init device properly not work
