+++ handlers/healthcheck.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+)
+
+func HealthCheckHandler(c echo.Context) error {
+	return c.String(http.StatusOK, "OK")
+}
