+++ services/devices.go
@@ -1,0 +1,9 @@
+package services
+
+import "github.com/labstack/echo/v4"
+
+type (
+	DevicesService interface {
+		InitDevice(c echo.Context) error
+	}
+)
