+++ views/index.html
@@ -21,7 +21,7 @@
 
 
 {{ block "form" . }}
-<form hx-targe="#form" hx-post="/save" class="form-settings">
+<form hx-targe="#form" hx-post="/api/v1/settings/save" class="form-settings">
     <!-- add range brightness -->
     <label for="brightness" class="input-base">
         <span>brightness</span>
